chore: refactored new-request-form module folder structure
