+++ src/modules/new-request-form/index.tsx
@@ -1,0 +1,1 @@
+export { renderNewRequestForm } from "./renderNewRequestForm";
