+++ src/modules/new-request-form/renderNewRequestForm.tsx
@@ -1,7 +1,7 @@
 import { render } from "react-dom";
+import { ComponentProviders } from "../shared";
 import type { NewRequestFormProps } from "./NewRequestForm";
 import { NewRequestForm } from "./NewRequestForm";
-import { ComponentProviders } from "../shared";
 
 export function renderNewRequestForm(
   props: NewRequestFormProps,
